Add Personas
Problem
-------
Roles are often not fine grained enough, since the "role" a person
has can change depending on the state of a workflow.

For example, suppose there are three users, Alice, Ben, and Carla.

Alice is a supervisor, and assigns at task to Ben. Ben is responsible
to Alice for completing the task, but he has the option to assign it
to Carla. If he *does* assign it, he's able to do certain actions
(like review Carla's work).

Solution
--------
Provide "personas", which are more flexible than roles. Just as
the persona originated in classic theatre, where an actor would
take on one mask, the another to play a different role, so here.