Add Rules to Limn
Problem
-------
Often we have business rules that need to be converted to facts.

For example, if we are making coffee and need a minimum of three scoops,
then we might need a fact `:sufficient-coffee-grounds`.

Solution
--------
Add a `workflow/rules` map of functions that can be applied to the facts
to derive new facts.