Implement an orchestrator
Problem
-------
`limn` currently answers the question of who does what when in the
context of a workflow. But it leaves the data fetching and events up to
the caller.

However, this does not fully encapsulate domain logic.

Solution
--------
Create an orchestrator for a functional core, imperative shell approach.
The orchestrator lets the core declare its data needs, orchestrates
fetching the data, then calls the core again for its decisions as to
what effects to carry out.