Connect limn orchestrator with limn workflow logic
Problem
-------
Limn's workflows solve the problem of "guards" - is this a permitted
action in the context of a workflow, given the state.

Limn's orchestrator can delegate this work to the `decider` function.
But that is extra plumbing, and requires the decider to include guards.

Solution
--------
Limn's orchestrator will, when given an `action` (a map with an
`:action/name`) and a workflow, check with the workflow to see if the
action is allowed. If not, it returns an anomaly.